consts refactored
