+++ src/p2p/consts.js
@@ -1,0 +1,9 @@
+module.exports = {
+	IPFS_WRITE_OPTS: {
+		create: true,
+		parents: true,
+		truncate: true
+	},
+	IPFS_MSG_PATH: '/msgs',
+	RESPONSE_TIMEOUT: 10000,
+}
